Update auth models

+++ authapi-service/src/core/out_models/auth.py
@@ -3,6 +3,6 @@ from src.core.out_models.base import Base
 
 class SuccessfulLogin(Base):
     access_token: str
-    access_token_expire_timestamp: float
+    access_token_expire_timestamp: int
     refresh_token: str
-    refresh_token_expire_timestamp: float
+    refresh_token_expire_timestamp: int
